Implement PKCS#7 envelope parsing
This additionally implements a pending challenge table within our database, which is inserted upon enrollment.

As a response to enrollment, the device POSTs a PKCS#7 envelope to the specified enrollment endpoint. Initially, it is signed by Apple's iPhone Device CA, and will utilize the certificate we issue it later on.

This commit is quite large, and handling of contents within enrollment is incomplete. This will be completed later on.